Fix text image gen when autofit: false In this case the text image is generated using SVG since this gives more control over stroke and fill attributes.  However it also has limitations generating images of consistent font size when the image width is calculated from the text string (rather than a defined width as is the case with autofit: true).
This commit changes strategy to calculate an estimated
final image width, generates the text image then
trims the image to the size of the generated text.

As a result, the text in the generated images is of
a consistent size for a given font_size. However the
generated image will have diffierent heights depending
on the string rendered based upon whether the string
has majiscules or miniscules.

Thanks to @ilesar for the report. Closes #86.