use file_ref instead of unused binary field from `Kino.Input.read/1`
The previous return of `Kino.Input.read/1` was a map with one of the keys `binary` which value was the binary of the image. This has changed to a field `file_ref` which is only a reference to a file which can be read with `Kino.Input.file_path/1`. See https://hexdocs.pm/kino/Kino.Input.html#file_path/1